refactor: 리포지토리 및 테스트 코드 리팩토링
1. SpringDataColorRepository와 ColorRepository 분리 구조 적용한다.
2. Color 엔티티의 정적 팩토리 메서드(Color.create)로 코드 단순화한다.
3. 불필요한 테스트 코드 주석 처리 및 수정한다.

Resolves #34
